backend/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-filter==25.1
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
whitenoise==6.6.0
pgvector==0.4.1 

# Scraping and web utilities
--find-links=./wheels
setuptools>=65.5.0
instaloader==4.14.2+custom1
python-dotenv
python-dateutil==2.9.0.post0
requests==2.31.0
beautifulsoup4
openai

# Email service
resend==0.8.0

# Encryption for newsletter emails
cryptography==41.0.7

# AWS S3 for image storage
boto3==1.34.0
Pillow>=10.0.0

# Development and deployment
gunicorn==21.2.0 